do some changes in backend

+++ springboot-ticketing-service/src/main/java/com/Java_CW/TicketBooking/service/TicketPoolInterface.java
@@ -14,5 +14,5 @@ public interface TicketPoolInterface {
 	 * @param customerId
 	 * @param customerRetrievalRate
 	 */
-	void removeTicket(int customerId, double customerRetrievalRate);
+	void removeTicket(int customerId, int customerRetrievalRate);
 }
